requirements.txt
streamlit
azure-identity==1.25.2
azure-ai-projects==1.0.0
python-dotenv